Added a loading gif while voting.
I could not stop thinking about this feature. Had to code it before I
could study in peace. >.<
When the user clicks to vote, a loader gif is shown beside the
position, while he is ready to vote for the next person. Only when the
browser receives confirmation of the vote being received by the server,
will the loading gif turn into a tick mark.